Refactors the code and adds a first class with the Package Management commands
